[********]<feature>核间负载数据解析功能前后端通信部分+单元测试+检视意见修改
Match-id-6e7d14ab6433f5920a9970045dfd921359eb492e

+++ server/src/modules/source/parser/InterCoreLoadGraphParser.cpp
@@ -122,7 +122,7 @@ void InterCoreLoadGraphParser::ParseJsonOpDetailArray(InterCoreLoadAnalysisDetai
             subCoreDetail.subCoreIndex = JsonUtil::GetInteger(jsonCoreDetail, "subcore_id");
             subCoreDetail.subCoreType = JsonUtil::GetString(jsonCoreDetail, "subcore_type");
             subCoreDetail.cycles = JsonUtil::GetInteger(jsonCoreDetail, "cycles");
-            subCoreDetail.throughput = JsonUtil::GetFloat(jsonCoreDetail, "throughput");
+            subCoreDetail.throughput = JsonUtil::GetInteger(jsonCoreDetail, "throughput");
             subCoreDetail.hitRate = JsonUtil::GetFloat(jsonCoreDetail, "L2cache_hit_rate");
 
             // 设置最优值
